Fix the use of assembly files
automake does not add a --tag switch to its libtool invocation when assembling
a .s file and rely on libtool to infer the right action based on the compiler
name. As shave is using CC to hook a wrapper, libtool gets confused. Let's
detect these cases and add a --tag=CC option.

Stefan Kost reported the bug.