Fix error caused by quoted arguments with spaces.
If an argument containing a space character was passed to the target
tool, shave would not interpret the quotes correctly, causing the tool
to interpret it as two different arguments.

When AM_CONFIG_HEADER is not used, a PACKAGE_STRING is passed to gcc
through the command line, and the value of this string is the package
name and package version seperated by a space. So in that case, shave
would not work.